bug fixes and updates

+++ app/src/main/java/com/cabalry/util/MovingAverage.java
@@ -5,19 +5,19 @@ package com.cabalry.util;
  */
 public class MovingAverage {
 
-    private int[] mSamples;
-    private int mSum;
+    private double[] mSamples;
+    private double mSum;
     private int mCurrentSample;
     private int mSampleCount;
 
     public MovingAverage(int size) {
-        mSamples = new int[size];
+        mSamples = new double[size];
         mSum = 0;
         mCurrentSample = 0;
         mSampleCount = 0;
     }
 
-    public void addSample(int sample) {
+    public void addSample(double sample) {
         if (mSampleCount == mSamples.length) {
             mSum -= mSamples[mCurrentSample];
         } else {
@@ -33,7 +33,7 @@ public class MovingAverage {
         }
     }
 
-    public int getCurrentAverage() {
+    public double getCurrentAverage() {
         if (mSampleCount != 0) {
             return mSum / mSampleCount;
         }
